Added third .sh file and fixed a small bug in MapLoader.h
The third .sh file does all git magic and lets build.sh only build. With this construction the
lines cleaning the directory doesn't need to be commented away when doing local debugging. This
is instead done in the new .sh file. When using templates care muste be taken to place a
whitespace between the brackets when nestling them.This is now fixed in MapLoader.h