**PLEASE READ!!** Implemented easiest changes possible to enable state changes. STRUCTURAL CHANGES IN THESE COMMIT FILES HAVE BEEN NOTED. OLD ALLOC/DEALLOC PLACES ARE NOT DELETED BUT COMMENTED OUT. IF THIS SYSTEM IS MERGED WITH MASTER THESE SHOULD BE PROPERLY DELETED.
It is built upon the previous system of allocation/deallocation of resources.
It has been tweaked to be able to do this in a controlled manner and in the same places.

This is now handled by onEntry and onExit methods in the states. Every time a state
is exited it deallocates its resources. And every time it is entered it allocates them(this is not done in the constructor).

The StateManager calls onExit on m_currentState in switch right before the switch is performed.

A system utilising pooling instead can be implemented upon this architecture now. Though it probably requires at least some structural change on the state base class.