The menu now supports scaling. TextArea now has anchor points. Started to write tests.
The menu now correctly supports scaling. All positioning and sizing of menu items are now done in the range 0.0 - 1.0. You can now bind text to be rendered top/center/bottom and left/center/right of the origin position. Tests are not complete. They need a lot of more work and thinking in how they should be done.