Basic functionality for debyugging OGl added.
A function that calls glGetError() until all errors are handled is implemented.
This function also prints the errors in the console. This function never ends
as it's using a while loop that breaks when the error que is empty, which it
never is. when using modern NVIDIA drivers. Testing need to be done on older
drivers and ATI cards.

A OGL debug class has been added. This is used to translate the error codes to
strings. The error fetching funtionality is broken due to some linking errors
when used.