Resolved bug.
Parts of InGameState were not initialized when restart() was called upon. The
reason for this is that states are deallocated when they are not active. The
bug was resolved by letting the updateloop look for requests for map change and
then calling restart from update. This ensures that the state is in fact
active.